requirements.txt
Django>=4.0
djangorestframework
transformers
torch
gunicorn